% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tbl-sql-all.r
\name{build_query}
\alias{build_query}
\title{Build query}
\usage{
build_query(x, limit = NULL)
}
\arguments{
\item{x}{Object for which query is to be build.}

\item{limit}{Limit to apply}
}
\description{
Build query
}

