% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translate-sql-base.R
\name{copy_env}
\alias{copy_env}
\title{Copy Environment}
\usage{
copy_env(from, to = NULL, parent = parent.env(from))
}
\arguments{
\item{from}{Envrironment to copy from}

\item{to}{Environment to copy to}

\item{parent}{Paren environment of copied elements}
}
\description{
Copy Environment
}

