% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-oracle.r
\name{copy_to.src_oracle}
\alias{copy_to.src_oracle}
\title{Copy to oracle}
\usage{
\method{copy_to}{src_oracle}(dest, df, name = deparse(substitute(df)),
  types = NULL, temporary = FALSE, indexes = NULL, analyze = TRUE, ...)
}
\arguments{
\item{dest}{Destination}

\item{df}{Data frame to copy}

\item{name}{Name in database}

\item{types}{Types}

\item{temporary}{Should table be temporary?}

\item{indexes}{Table indexes}

\item{analyze}{Should table statistics be recalculated after copy?}

\item{...}{Other arguments passed to function.}
}
\description{
Copy to oracle
}

