% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customSqlFunctions.R
\name{db_drop_tables}
\alias{db_drop_tables}
\title{Drop multiple tables in Oracle database}
\usage{
db_drop_tables(con, tables, force = FALSE, ...)
}
\arguments{
\item{con}{Connection}

\item{tables}{Character vector with table names to drop.}

\item{force}{Force drop?}

\item{...}{Other arguments passed to function}
}
\description{
Drop multiple tables in Oracle database
}

