% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-oracle.r
\name{\%||\%}
\alias{\%||\%}
\title{#' @export
sql_subquery.OraConnection <- function(con, sql, name = dplyr:::unique_name(), ...) {
 if (is.ident(sql)) return(sql)
 build_sql("(", sql, ") ", ident(name), con = con)
}
#' @export
sql_subquery.OraConnection <- function(con, from, name = unique_name(), ...) {
  if (is.ident(from)) {
    setNames(from, name)
  } else {
    if (is.null(name)) {
      build_sql("(", from, ")", con = con)
    } else {
      build_sql("(", from, ") AS ", ident(name), con = con)
    }
  }
}}
\usage{
x \%||\% y
}
\description{
#' @export
sql_subquery.OraConnection <- function(con, sql, name = dplyr:::unique_name(), ...) {
 if (is.ident(sql)) return(sql)
 build_sql("(", sql, ") ", ident(name), con = con)
}
#' @export
sql_subquery.OraConnection <- function(con, from, name = unique_name(), ...) {
  if (is.ident(from)) {
    setNames(from, name)
  } else {
    if (is.null(name)) {
      build_sql("(", from, ")", con = con)
    } else {
      build_sql("(", from, ") AS ", ident(name), con = con)
    }
  }
}
}

