% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lahman.r
\name{lahman_oracle}
\alias{lahman_oracle}
\title{Cache and retrieve an \code{src_oracle} of the Lahman baseball database.}
\usage{
lahman_oracle(src, ...)
}
\arguments{
\item{src}{Source to use.}

\item{...}{Other arguments passed to function}
}
\description{
This creates an interesting database using data from the Lahman baseball
data source, provided by Sean Lahman at
\url{http://www.seanlahman.com/baseball-archive/statistics/}, and
made easily available in R through the \pkg{Lahman} package by
Michael Friendly, Dennis Murphy and Martin Monkman. See the documentation
for that package for documentation of the inidividual tables.
}

