% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{ora_sql}
\alias{ora_sql}
\title{ora_sql}
\usage{
ora_sql(query, tolower = TRUE, dots = TRUE, encoding = "unknown",
  useBytes = TRUE, stringsAsFactors = FALSE, warn = -1, debug = FALSE,
  ...)
}
\arguments{
\item{query}{xxx}

\item{tolower}{xxx}

\item{dots}{xxx}

\item{encoding}{xxx}

\item{useBytes}{xxx}

\item{stringsAsFactors}{xxx}

\item{warn}{xxx}

\item{debug}{xxx}

\item{...}{xxx}
}
\description{
ora_sql
}

