% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{oracle_tables}
\alias{oracle_tables}
\title{oracle_tables}
\usage{
oracle_tables(owner = "\%", table = "\%", space = "\%", tolower = TRUE,
  ...)
}
\arguments{
\item{owner}{xxx}

\item{table}{xxx}

\item{space}{xxx}

\item{tolower}{xxx}

\item{...}{xxx}
}
\description{
oracle_tables
}

