% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{oracle_variables}
\alias{oracle_variables}
\title{oracle_variables}
\usage{
oracle_variables(table, tolower = TRUE, dots = FALSE, ...)
}
\arguments{
\item{table}{xxx}

\item{tolower}{xxx}

\item{dots}{xxx}

\item{...}{xxx}
}
\description{
oracle_variables
}

