% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/over.R
\name{over}
\alias{over}
\title{Set of internal over functions}
\usage{
over(expr, partition = NULL, order = NULL, frame = NULL)
}
\description{
Set of internal over functions
}
\keyword{NULL}
\keyword{Over}
\keyword{internal}

