% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-base.R
\name{sql_translator}
\alias{sql_translator}
\title{Set of SQL Translator functtions}
\usage{
sql_translator(..., .funs = list(), .parent = new.env(parent = emptyenv()))
}
\description{
Set of SQL Translator functtions
}
\keyword{NULL}
\keyword{SQL}
\keyword{Translator}
\keyword{internal}

