% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/over.R
\name{sql_vector_over}
\alias{sql_vector_over}
\title{SQL Vector Over
Collapse arguments in such way it can be used in partition/order by
clause of window functions}
\usage{
sql_vector_over(x, parens = NA, collapse = " ", con = NULL)
}
\arguments{
\item{x}{xx}

\item{parens}{xx}

\item{collapse}{xx}

\item{con}{xx}
}
\description{
SQL Vector Over
Collapse arguments in such way it can be used in partition/order by
clause of window functions
}

