% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-oracle.r
\name{src_oracle}
\alias{src_oracle}
\title{Connect to oracle.}
\usage{
src_oracle(dbname = NULL, host = NULL, port = NULL, user = NULL,
  password = NULL, ...)
}
\arguments{
\item{dbname}{Oracle database name}

\item{host}{Host name}

\item{port}{Port}

\item{user}{User name}

\item{password}{Password}

\item{...}{Other arguments passed to function.}
}
\description{
Connect to oracle.
}
\details{
Note that you can also use SID based connection if supported by
your infrastructure
}
\examples{
\dontrun{
host <- "11.111.11.11"
port <- 1521
sid <- "mydbident"

connect.string <- paste(
 "(DESCRIPTION=",
 "(ADDRESS=(PROTOCOL=tcp)(HOST=", host, ")(PORT=", port, "))",
 "(CONNECT_DATA=(SID=", sid, ")))", sep = "")

my_db <- src_oracle(username = "ilikedb", password = "verysecret", 
                   dbname = connect.string)
}
}

