% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customSqlFunctions.R
\name{tbls}
\alias{tbls}
\title{Create tables in global environment}
\usage{
tbls(src, ..., tables, envir = globalenv(), tolower = TRUE)
}
\arguments{
\item{src}{Source}

\item{...}{Other arguments passed to function}

\item{tables}{Character vector with table names.}

\item{envir}{Environment where tables are created.}

\item{tolower}{Should variables pointing to tables be lowercased?}
}
\description{
Create tables in global environment
}

