% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/customSqlFunctions.R
\name{union_all}
\alias{union_all}
\title{Union all operator}
\usage{
union_all(x, y, ...)
}
\arguments{
\item{x}{Table \code{x} to union.}

\item{y}{Table \code{y} to union.}

\item{...}{Other arguments passed to function}
}
\description{
Union all operator
}

