% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/customSqlFunctions.R
\name{union_all.tbl_oracle}
\alias{union_all.tbl_oracle}
\title{Union all operator for Oracle}
\usage{
\method{union_all}{tbl_oracle}(x, y, copy = FALSE, ...)
}
\arguments{
\item{x}{Table \code{x} to union.}

\item{y}{Table \code{y} to union.}

\item{copy}{Should data be copied to source if not present?}

\item{...}{Other arguments passed to function}
}
\description{
Union all operator for Oracle
}

