% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-sql-all.r
\name{update.tbl_oracle}
\alias{update.tbl_oracle}
\title{Update}
\usage{
\method{update}{tbl_oracle}(object, ...)
}
\arguments{
\item{object}{Object to update.}

\item{...}{Other functions passed to function.}
}
\description{
Update
}

