% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translate-sql-base.R
\name{win_rank}
\alias{win_rank}
\title{Win Rank
I had to create a copy of win_cumulative in order to use my own
sql_vector function}
\usage{
win_rank(f)
}
\arguments{
\item{f}{Function to create}
}
\description{
Win Rank
I had to create a copy of win_cumulative in order to use my own
sql_vector function
}

