% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-sql-base.R
\name{win_recycled}
\alias{win_recycled}
\title{Win recycled Oracle
I had to create a copy of win_cumulative in order to use my own
sql_vector function}
\usage{
win_recycled(f)
}
\arguments{
\item{f}{Function to create}
}
\description{
Win recycled Oracle
I had to create a copy of win_cumulative in order to use my own
sql_vector function
}

